SOA/A/requirements.txt
Django==5.1.4
psycopg2==2.9.10
psycopg-binary==3.2.3
django-phonenumber-field==8.0.0
pillow==11.0.0
django-graphql-jwt==0.4.0
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
pillow==11.0.0
graphene-file-upload==1.3.0